Change to use assembly plugin for version revving
