+++ metron-deployment/packaging/ambari/metron-mpack/src/main/resources/addon-services/METRON/CURRENT/metainfo.xml
@@ -19,8 +19,8 @@
     <services>
         <service>
             <name>METRON</name>
-            <version>0.3.0</version>
-            <extends>common-services/METRON/0.3.0</extends>
+            <version>${project.parent.version}</version>
+            <extends>common-services/METRON/${project.parent.version}</extends>
         </service>
     </services>
 </metainfo>
