+++ metron-deployment/packaging/ambari/metron-mpack/src/main/resources/common-services/METRON/CURRENT/configuration/metron-env.xml
@@ -20,7 +20,7 @@
 <configuration supports_final="true">
     <property>
         <name>metron_home</name>
-        <value>/usr/metron/0.3.0</value>
+        <value>/usr/metron/${project.parent.version}</value>
         <description>Metron home directory</description>
         <display-name>Metron home</display-name>
     </property>
