+++ metron-deployment/packaging/ambari/metron-mpack/src/main/resources/common-services/METRON/CURRENT/package/scripts/indexing_commands.py
@@ -20,63 +20,54 @@ import time
 
 from resource_management.core.logger import Logger
 from resource_management.core.resources.system import Execute, File
-from resource_management.libraries.resources.hdfs_resource import HdfsResource
 
 import metron_service
 
 
 # Wrap major operations and functionality in this class
-class EnrichmentCommands:
+class IndexingCommands:
     __params = None
-    __enrichment_topology = None
-    __enrichment_topic = None
+    __indexing = None
     __configured = False
 
     def __init__(self, params):
         if params is None:
             raise ValueError("params argument is required for initialization")
         self.__params = params
-        self.__enrichment_topology = params.metron_enrichment_topology
-        self.__enrichment_topic = params.metron_enrichment_topic
-        self.__configured = os.path.isfile(self.__params.enrichment_configured_flag_file)
+        self.__indexing = params.metron_indexing_topology
+        self.__configured = os.path.isfile(self.__params.indexing_configured_flag_file)
 
     def is_configured(self):
         return self.__configured
 
     def set_configured(self):
-        File(self.__params.enrichment_configured_flag_file,
+        File(self.__params.indexing_configured_flag_file,
              content="",
              owner=self.__params.metron_user,
              mode=0775)
 
     def setup_repo(self):
-
         def local_repo():
             Logger.info("Setting up local repo")
             Execute("yum -y install createrepo")
             Execute("createrepo /localrepo")
             Execute("chmod -R o-w+r /localrepo")
+            Execute("echo \"[METRON-${project.parent.version}]\n"
+                    "name=Metron ${project.parent.version} packages\n"
+                    "baseurl=file:///localrepo\n"
+                    "gpgcheck=0\n"
+                    "enabled=1\" > /etc/yum.repos.d/local.repo")
 
         def remote_repo():
-            Logger.info('Using remote repo')
+            print('Using remote repo')
 
         yum_repo_types = {
             'local': local_repo,
             'remote': remote_repo
         }
-
         repo_type = self.__params.yum_repo_type
-
         if repo_type in yum_repo_types:
             yum_repo_types[repo_type]()
-            Logger.info("Writing out repo file")
-            repo_template = ("echo \"[METRON-0.3.0]\n"
-                            "name=Metron 0.3.0 packages\n"
-                            "baseurl={0}\n"
-                            "gpgcheck=0\n"
-                            "enabled=1\n\""
-                         "   > /etc/yum.repos.d/metron.repo")
-            Execute(repo_template.format(self.__params.repo_url))
         else:
             raise ValueError("Unsupported repo type '{0}'".format(repo_type))
 
@@ -93,85 +84,59 @@ class EnrichmentCommands:
         replication_factor = 1
         retention_gigabytes = int(self.__params.metron_topic_retention)
         retention_bytes = retention_gigabytes * 1024 * 1024 * 1024
-        Logger.info("Creating topics for enrichment")
+        Logger.info("Creating topics for indexing")
 
-        Logger.info("Creating topic'{0}'".format(self.__enrichment_topic))
+        Logger.info("Creating topic'{0}'".format(self.__indexing))
         Execute(command_template.format(self.__params.kafka_bin_dir,
                                         self.__params.zookeeper_quorum,
-                                        self.__enrichment_topic,
+                                        self.__indexing,
                                         num_partitions,
                                         replication_factor,
                                         retention_bytes))
         Logger.info("Done creating Kafka topics")
 
-    def init_hdfs_dir(self):
-        self.__params.HdfsResource(self.__params.metron_apps_enrichment_dir,
-                                   type="directory",
-                                   action="create_on_execute",
-                                   owner=self.__params.metron_user,
-                                   group=self.__params.user_group,
-                                   mode=0775,
-                                   )
-
-    def start_enrichment_topology(self):
-        Logger.info("Starting Metron enrichment topology: {0}".format(self.__enrichment_topology))
-        start_cmd_template = """{0}/bin/start_enrichment_topology.sh \
+    def start_indexing_topology(self):
+        Logger.info("Starting Metron indexing topology: {0}".format(self.__indexing))
+        start_cmd_template = """{0}/bin/start_elasticsearch_topology.sh \
                                     -s {1} \
                                     -z {2}"""
-        Logger.info('Starting ' + self.__enrichment_topology)
-        Execute(start_cmd_template.format(self.__params.metron_home, self.__enrichment_topology, self.__params.zookeeper_quorum))
+        Logger.info('Starting ' + self.__indexing)
+        Execute(start_cmd_template.format(self.__params.metron_home, self.__indexing, self.__params.zookeeper_quorum))
 
-        Logger.info('Finished starting enrichment topology')
+        Logger.info('Finished starting indexing topology')
 
-    def stop_enrichment_topology(self):
-        Logger.info('Stopping ' + self.__enrichment_topology)
-        stop_cmd = 'storm kill ' + self.__enrichment_topology
+    def stop_indexing_topology(self):
+        Logger.info('Stopping ' + self.__indexing)
+        stop_cmd = 'storm kill ' + self.__indexing
         Execute(stop_cmd)
-        Logger.info('Done stopping enrichment topologies')
+        Logger.info('Done stopping indexing topologies')
 
-    def restart_enrichment_topology(self, env):
-        Logger.info('Restarting the enrichment topologies')
-        self.stop_enrichment_topology()
+    def restart_indexing_topology(self, env):
+        Logger.info('Restarting the indexing topologies')
+        self.stop_indexing_topology()
 
         # Wait for old topology to be cleaned up by Storm, before starting again.
         retries = 0
         topology_active = self.is_topology_active(env)
-        while topology_active and retries < 3:
+        while self.is_topology_active(env) and retries < 3:
             Logger.info('Existing topology still active. Will wait and retry')
-            time.sleep(40)
-            topology_active = self.is_topology_active(env)
+            time.sleep(10)
             retries += 1
 
         if not topology_active:
-            self.start_enrichment_topology()
-            Logger.info('Done restarting the enrichment topology')
+            Logger.info('Waiting for storm kill to complete')
+            time.sleep(30)
+            self.start_indexing_topology()
+            Logger.info('Done restarting the indexing topologies')
         else:
             Logger.warning('Retries exhausted. Existing topology not cleaned up.  Aborting topology start.')
 
     def is_topology_active(self, env):
         env.set_params(self.__params)
-
         active = True
         topologies = metron_service.get_running_topologies()
         is_running = False
-        if self.__enrichment_topology in topologies:
-            is_running = topologies[self.__enrichment_topology] in ['ACTIVE', 'REBALANCING']
+        if self.__indexing in topologies:
+            is_running = topologies[self.__indexing] in ['ACTIVE', 'REBALANCING']
         active &= is_running
         return active
-
-    def create_hbase_tables(self):
-        add_enrichment_cmd = "echo \"create '{0}','{1}'\" | hbase shell -n".format(self.__params.enrichment_table, self.__params.enrichment_cf)
-        Execute(add_enrichment_cmd,
-                tries=3,
-                try_sleep=5,
-                logoutput=False,
-                path='/usr/sbin:/sbin:/usr/local/bin:/bin:/usr/bin'
-                )
-
-        add_threatintel_cmd = "echo \"create '{0}','{1}'\" | hbase shell -n".format(self.__params.threatintel_table, self.__params.threatintel_cf)
-        Execute(add_threatintel_cmd,
-                tries=3,
-                try_sleep=5,
-                logoutput=False,
-                path='/usr/sbin:/sbin:/usr/local/bin:/bin:/usr/bin'
-                )
