+++ metron-deployment/packaging/ambari/metron-mpack/src/main/resources/common-services/METRON/CURRENT/package/scripts/parser_commands.py
@@ -77,8 +77,8 @@ class ParserCommands:
             Execute("yum -y install createrepo")
             Execute("createrepo /localrepo")
             Execute("chmod -R o-w+r /localrepo")
-            Execute("echo \"[METRON-0.3.0]\n"
-                    "name=Metron 0.3.0 packages\n"
+            Execute("echo \"[METRON-${project.parent.version}]\n"
+                    "name=Metron ${project.parent.version} packages\n"
                     "baseurl=file:///localrepo\n"
                     "gpgcheck=0\n"
                     "enabled=1\" > /etc/yum.repos.d/local.repo")
